Move each place component into separate files.
