+++ static/js/ranking/ranking_page.tsx
@@ -20,7 +20,7 @@ import { STATS_VAR_TITLES } from "../shared/stats_var_titles";
 import { LocationRankData } from "./ranking_types";
 import { RankingHistogram } from "./ranking_histogram";
 import { RankingTable } from "./ranking_table";
-import { displayNameForPlaceType } from "../shared/util";
+import { displayNameForPlaceType } from "../place/place_util";
 
 const GET_BOTTOM_PARAM = "bottom";
 const RANK_SIZE = 100;
